fix: handle empty text/caption in replies
